// backend/template/src/app/api/model/voteDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VoteDTO { 
    id?: number;
    votableId?: number;
    voteType?: VoteDTO.VoteTypeEnum;
    userId?: number;
}
export namespace VoteDTO {
    export type VoteTypeEnum = 'DOWNVOTE' | 'UPVOTE';
    export const VoteTypeEnum = {
        Downvote: 'DOWNVOTE' as VoteTypeEnum,
        Upvote: 'UPVOTE' as VoteTypeEnum
    };
}